pywww/requirements.txt
django==3.2.9
django-extensions==3.1.5
django-import-export
PyYAML==6.0
Faker==9.5.2
pillow
django-crispy-forms